Development v0.12.000

+++ RobotComponents.Gh/Components/Deconstruct/DeconstructMovementComponent.cs
@@ -76,7 +76,7 @@ namespace RobotComponents.Gh.Components.Deconstruct
             // Output
             DA.SetData(0, movement.Target);
             DA.SetData(1, movement.SpeedData);
-            DA.SetData(2, movement.MovementType);
+            DA.SetData(2, (int)movement.MovementType);
             DA.SetData(3, movement.ZoneData);
             DA.SetData(4, movement.RobotTool);
             DA.SetData(5, movement.WorkObject);
